feat: Refactor and modularize utility functions for URL validation and format sorting
